Revert changes to http-client
